sensor_simulator/requirements.txt
# mqtt_workshop/sensor_simulator/requirements.txt
paho-mqtt